+++ language-service/javascript/src/index.ts
@@ -1,4 +1,5 @@
 export * from './getGherkinSemanticTokens'
 export * from './getGherkinDiagnostics'
 export * from './getGherkinCompletionItems'
+export * from './getGherkinFormattingEdits'
 export * from './CucumberInfo'